requirements.txt
absl-py==1.1.0
ansi2html==1.7.0
asgiref==3.5.2
attrs==21.4.0
Brotli==1.0.9
brotlipy==0.7.0
certifi==2021.10.8
click==8.1.3
cycler==0.11.0
dj-database-url==1.0.0
Django==4.0.5
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-markdownx==3.0.1
django-on-heroku==1.1.2
Markdown==3.3.7
openpyxl==3.0.9
Pillow==9.1.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cpu
pymysql==1.0.2
mysqlclient==2.0.3
captcha==0.3
numpy==1.21.4
pandas==1.3.4
scipy==1.7.3